+++ feature/mybookdetail/src/main/java/app/kaito_dogi/mybrary/feature/mybookdetail/MyBookDetailNavArg.kt
@@ -5,10 +5,10 @@ import app.kaito_dogi.mybrary.core.domain.model.Memo
 import app.kaito_dogi.mybrary.core.domain.model.MemoId
 import app.kaito_dogi.mybrary.core.domain.model.MyBook
 import app.kaito_dogi.mybrary.core.domain.model.MyBookId
-import app.kaito_dogi.mybrary.core.ui.MemoIdParceler
-import app.kaito_dogi.mybrary.core.ui.MemoParceler
-import app.kaito_dogi.mybrary.core.ui.MyBookIdParceler
-import app.kaito_dogi.mybrary.core.ui.MyBookParceler
+import app.kaito_dogi.mybrary.core.ui.parceler.MemoIdParceler
+import app.kaito_dogi.mybrary.core.ui.parceler.MemoParceler
+import app.kaito_dogi.mybrary.core.ui.parceler.MyBookIdParceler
+import app.kaito_dogi.mybrary.core.ui.parceler.MyBookParceler
 import kotlinx.parcelize.Parcelize
 import kotlinx.parcelize.TypeParceler
 import kotlinx.serialization.Serializable
